Merged 7aea838 from https://github.com/CTSRD-CHERI/v8/commit/7aea838c14c51d4968afb9048da9d219156c79cb

+++ src/base/pointer-with-payload.h
@@ -58,7 +58,7 @@ class PointerWithPayload {
   }
 
   V8_INLINE PointerType* GetPointer() const {
-    return reinterpret_cast<PointerType*>(pointer_with_payload_ & kPointerMask);
+    return reinterpret_cast<PointerType*>(pointer_with_payload_ & (size_t) kPointerMask);
   }
 
   // An optimized version of GetPointer for when we know the payload value.
